requirements.txt
numpy==2.1.3
pillow==11.0.0
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
open_clip_torch==2.26.1
torch>=2.2.0
torchvision>=0.17.0
tqdm==4.67.1
